+++ Assets/uNode3/Core.Editor/Window/AboutWindow.cs
@@ -3,8 +3,8 @@ using UnityEditor;
 
 namespace MaxyGames.UNode.Editors {
 	public class About : EditorWindow {
-		public const string version = "3.2.5";
-		public const float compilerVersion = 3.25f;
+		public const string version = "3.2.7";
+		public const float compilerVersion = 3.27f;
 
 		GUIStyle PublisherNameStyle, headerStyle, infoStyle;
 		GUIStyle ToolBarStyle;